+++ atlasdb-dbkvs/src/main/java/com/palantir/atlasdb/keyvalue/dbkvs/impl/OracleDbTableFactory.java
@@ -30,12 +30,14 @@ import com.palantir.nexus.db.DBType;
 
 public class OracleDbTableFactory implements DbTableFactory {
     private final OracleDdlConfig config;
-    private OracleTableNameGetter oracleTableNameGetter;
-    private TableValueStyleCache valueStyleCache;
+    private final OracleTableNameGetter oracleTableNameGetter;
+    private final OraclePrefixedTableNames oraclePrefixedTableNames;
+    private final TableValueStyleCache valueStyleCache;
 
     public OracleDbTableFactory(OracleDdlConfig config) {
         this.config = config;
         oracleTableNameGetter = new OracleTableNameGetter(config);
+        oraclePrefixedTableNames = new OraclePrefixedTableNames(oracleTableNameGetter);
         valueStyleCache = new TableValueStyleCache();
     }
 
@@ -95,7 +97,7 @@ public class OracleDbTableFactory implements DbTableFactory {
         TableValueStyle tableValueStyle = valueStyleCache.getTableType(conns, tableRef, config.metadataTable());
         switch (tableValueStyle) {
             case OVERFLOW:
-                return OracleOverflowWriteTable.create(config, conns, oracleTableNameGetter, tableRef);
+                return OracleOverflowWriteTable.create(config, conns, oracleTableNameGetter, oraclePrefixedTableNames, tableRef);
             case RAW:
                 return new OracleWriteTable(config, conns, oracleTableNameGetter, tableRef);
             default: