+++ atlasdb-dbkvs/src/main/java/com/palantir/atlasdb/keyvalue/dbkvs/impl/PrefixedTableNames.java
@@ -16,17 +16,9 @@
 package com.palantir.atlasdb.keyvalue.dbkvs.impl;
 
 import com.palantir.atlasdb.keyvalue.api.TableReference;
-import com.palantir.atlasdb.keyvalue.dbkvs.DdlConfig;
 
-public class PrefixedTableNames {
-    private DdlConfig config;
+public interface PrefixedTableNames {
 
-    public PrefixedTableNames(DdlConfig config) {
-        this.config = config;
-    }
-
-    public String get(TableReference tableRef) {
-        return config.tablePrefix() + DbKvs.internalTableName(tableRef);
-    }
+    String get(TableReference tableRef, ConnectionSupplier connectionSupplier);
 
 }