+++ atlasdb-dbkvs/src/main/java/com/palantir/atlasdb/keyvalue/dbkvs/impl/ranges/DbKvsGetRanges.java
@@ -42,6 +42,7 @@ import com.palantir.atlasdb.keyvalue.api.RangeRequests;
 import com.palantir.atlasdb.keyvalue.api.RowResult;
 import com.palantir.atlasdb.keyvalue.api.TableReference;
 import com.palantir.atlasdb.keyvalue.api.Value;
+import com.palantir.atlasdb.keyvalue.dbkvs.impl.ConnectionSupplier;
 import com.palantir.atlasdb.keyvalue.dbkvs.impl.DbKvs;
 import com.palantir.atlasdb.keyvalue.dbkvs.impl.PrefixedTableNames;
 import com.palantir.atlasdb.keyvalue.dbkvs.impl.oracle.PrimaryKeyConstraintNames;
@@ -192,23 +193,25 @@ public class DbKvsGetRanges {
         }
 
         String order = reverse ? "DESC" : "ASC";
-        if (numRowsToGet == 1) {
-            String minMax = reverse ? "max" : "min";
-            // QA-69854 Special case 1 row reads because oracle is terrible at optimizing queries
-            String query = dbType == DBType.ORACLE
-                    ? getSimpleRowSelectOneQueryOracle(tableRef, minMax, extraWhere)
-                    : getSimpleRowSelectOneQueryPostgres(tableRef, extraWhere, order);
-            return Pair.create(query, args);
-        } else {
-            String query = String.format(
-                    SIMPLE_ROW_SELECT_TEMPLATE,
-                    DbKvs.internalTableName(tableRef),
-                    getPrimaryKeyConstraintName(tableRef),
-                    getPrefixedTableName(tableRef),
-                    extraWhere,
-                    order);
-            String limitQuery = BasicSQLUtils.limitQuery(query, numRowsToGet, args, dbType);
-            return Pair.create(limitQuery, args);
+        try (ConnectionSupplier conns = new ConnectionSupplier(connectionSupplier)) {
+            if (numRowsToGet == 1) {
+                String minMax = reverse ? "max" : "min";
+                // QA-69854 Special case 1 row reads because oracle is terrible at optimizing queries
+                String query = dbType == DBType.ORACLE
+                        ? getSimpleRowSelectOneQueryOracle(tableRef, minMax, extraWhere, conns)
+                        : getSimpleRowSelectOneQueryPostgres(tableRef, extraWhere, order, conns);
+                return Pair.create(query, args);
+            } else {
+                    String query = String.format(
+                            SIMPLE_ROW_SELECT_TEMPLATE,
+                            DbKvs.internalTableName(tableRef),
+                            getPrimaryKeyConstraintName(tableRef, conns),
+                            getPrefixedTableName(tableRef, conns),
+                            extraWhere,
+                            order);
+                    String limitQuery = BasicSQLUtils.limitQuery(query, numRowsToGet, args, dbType);
+                    return Pair.create(limitQuery, args);
+            }
         }
     }
     /**
@@ -303,11 +306,12 @@ public class DbKvsGetRanges {
     private String getSimpleRowSelectOneQueryPostgres(
             TableReference tableRef,
             String extraWhere,
-            String order) {
+            String order,
+            ConnectionSupplier conns) {
         return String.format(
                 SIMPLE_ROW_SELECT_ONE_POSTGRES_TEMPLATE,
                 DbKvs.internalTableName(tableRef),
-                getPrefixedTableName(tableRef),
+                getPrefixedTableName(tableRef, conns),
                 extraWhere,
                 order);
     }
@@ -315,22 +319,23 @@ public class DbKvsGetRanges {
     private String getSimpleRowSelectOneQueryOracle(
             TableReference tableRef,
             String minMax,
-            String extraWhere) {
+            String extraWhere,
+            ConnectionSupplier conns) {
         return String.format(
                 SIMPLE_ROW_SELECT_ONE_ORACLE_TEMPLATE,
                 DbKvs.internalTableName(tableRef),
-                getPrimaryKeyConstraintName(tableRef),
+                getPrimaryKeyConstraintName(tableRef, conns),
                 minMax,
-                getPrefixedTableName(tableRef),
+                getPrefixedTableName(tableRef, conns),
                 extraWhere);
     }
 
-    private String getPrimaryKeyConstraintName(TableReference tableRef) {
-        return PrimaryKeyConstraintNames.get(getPrefixedTableName(tableRef));
+    private String getPrimaryKeyConstraintName(TableReference tableRef, ConnectionSupplier conns) {
+        return PrimaryKeyConstraintNames.get(getPrefixedTableName(tableRef, conns));
     }
 
-    private String getPrefixedTableName(TableReference tableRef) {
-        return prefixedTableNames.get(tableRef);
+    private String getPrefixedTableName(TableReference tableRef, ConnectionSupplier conns) {
+        return prefixedTableNames.get(tableRef, conns);
     }
 
     private static final String SIMPLE_ROW_SELECT_TEMPLATE =